backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Flask backend with MySQL (Docker port 13306)"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0"
flask-sqlalchemy = "^3.1"
flask-migrate = "^4.0"
flask-cors = "^4.0"
pymysql = "^1.1"
python-dotenv = "^1.0"
google-genai = "^1.56.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
